Changed primitives of Population from int to double and of RoundedValue from Double to String | Filled in computation worker functions

+++ src/main/java/de/itdesign/application/Population.java
@@ -2,14 +2,16 @@ package de.itdesign.application;
 
 public class Population {
 
-    private int population;
+    private Double population;
 
-    public int getPopulation() {
+    public Double getPopulation() {
         return population;
     }
 
-    public void setPopulation(int population) {
+    public void setPopulation(Double population) {
         this.population = population;
     }
+
+    
 
 }
